refactored collision Detection
