+++ characterEntities/Hero.java
@@ -1,8 +1,7 @@
 package characterEntities;
 
+import GameMaps.MapCollisionDetection;
 import animation.AbilityAnimation;
-import gui.DamageMarker;
-import screens.GameMap;
 
 import java.awt.*;
 import java.util.ArrayList;
@@ -47,8 +46,8 @@ public abstract class Hero extends Entity {
 	static final int DEFAULT_RANGE = SQUARE_LENGTH;
 
 	
-	Hero(ArrayList<Entity> targets, GameMap map, int maxHealth, int maxDamage, int minDamage, int posX, int posY, int velocity) {
-		super(map, maxHealth, maxDamage, minDamage, posX, posY, velocity);
+	Hero(ArrayList<Entity> targets, MapCollisionDetection mapCollisionDetection, int maxHealth, int maxDamage, int minDamage, int posX, int posY, int velocity) {
+		super(mapCollisionDetection, maxHealth, maxDamage, minDamage, posX, posY, velocity);
 		this.targets = targets;
 		for (Entity target : targets) {
 			immuneTo.put(target, false);
@@ -178,6 +177,6 @@ public abstract class Hero extends Entity {
 			}
 		}
 
-		setPoint(map.determineMotion(newPosX, newPosY, getEntitySize(), targets));
+		setPoint(mapCollisionDetection.determineMotion(newPosX, newPosY, getEntitySize(), targets));
 	}
 }
