+++ characterEntities/RedHero.java
@@ -1,13 +1,13 @@
 package characterEntities;
 
+import GameMaps.MapCollisionDetection;
 import gui.DamageMarker;
-import screens.GameMap;
 
 import java.util.ArrayList;
 
 public class RedHero extends Hero {
-	public RedHero(ArrayList<Entity> targets, GameMap map) {
-		super(targets, map, 30, 15, 5, 100, 100, 5);
+	public RedHero(ArrayList<Entity> targets, MapCollisionDetection mapCollisionDetection) {
+		super(targets, mapCollisionDetection, 30, 15, 5, 100, 100, 5);
 		setPlayerClass(PlayerClass.RED);
 		setImageIcon("src/assets/hero/redNeutral.png");
 	}
